% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_select.R
\name{col_select}
\alias{col_select}
\title{Function for showing a shiny gadget that allows you to select the desired columns in a data frame.}
\usage{
col_select(df, ret = c("df_select", "dplyr_code"), top_n = 100)
}
\arguments{
\item{df}{The data frame}
}
\description{
This function subsets a data frame based on columns you select in a gadget.
}
\examples{
col_select()

}
\keyword{dplyr}
\keyword{select}
