% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edist.R
\name{edist}
\alias{edist}
\title{Function for creating a numeric vector showing the smallest distance between the
polygon in the primary sf object and all polygons in the second object}
\usage{
edist(x, y)
}
\arguments{
\item{x}{the primary sf object}

\item{y}{the second sf object}
}
\description{
Function for creating a numeric vector showing the smallest distance between the
polygon in the primary sf object and all polygons in the second object
}
\examples{
records <- read.xlsx("records.xlsx")
lws <- st_read("G:/Data/Data to Use/Sites/Berks/LWS/GIS data/2019/Berkshire Local Wildlife Sites April 2019.TAB")
records <- OS2POLY(records)
records$distance <- edist(records, lws)

}
\keyword{distance,}
\keyword{eucledian}
\keyword{polygons}
\keyword{sf,}
\keyword{spatial,}
