% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EN2OS.R
\name{en2os}
\alias{en2os}
\title{Eastings and Northings to OS Grid Ref function}
\usage{
en2os(easting, northing, resolution)
}
\arguments{
\item{easting}{The eastings of the point to be converted.}

\item{northing}{The northings of the point to be converted.}

\item{resolution}{The resolution of the grid reference to be produced.}
}
\description{
This function converts eastings and northings to an OS grid ref of a specified length.
}
\examples{
en2os()
}
\keyword{Grid}
\keyword{references}
