% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_png.R
\name{get_png}
\alias{get_png}
\title{Function for reading a png as a raster}
\usage{
get_png(filename)
}
\arguments{
\item{filename}{the  filename of the png file}
}
\description{
Function for reading a png as a raster
}
\examples{
library(sf)
logo <- get_png(system.file("data/TVERC_logo.png", package = "ALERC"))
myshape <- st_read(system.file("data/BerksOxon.shp", package = "ALERC"))
library(ggplot2)
ggplot(myshape) +
  geom_sf() +
  coord_sf(datum = 27700) +
  annotation_custom(logo, xmin = 470000, xmax = 504000, ymin = 240000, ymax = Inf) +
  theme_tverc()
}
\keyword{ggplot}
\keyword{mapping,}
\keyword{png,}
