% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatch.R
\name{hatch}
\alias{hatch}
\title{Function for creating an sf lines object to resemble hatching within a polygon for plotting on a map
This function returns an sf linestring or multilinestring object.}
\usage{
hatch(x, pattern, bbox)
}
\arguments{
\item{x}{the  sf object containing polygons that you want hatching}

\item{pattern}{horizontal, vertical, left2right, right2left - determines the hatching angle}
}
\description{
Function for creating an sf lines object to resemble hatching within a polygon for plotting on a map
This function returns an sf linestring or multilinestring object.
}
\keyword{hatch,}
\keyword{mapping}
\keyword{sf,}
