% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movecols.R
\name{movecols}
\alias{movecols}
\title{Remove rows based on NA values in one column}
\usage{
movecols(invec, movecommand)
}
\description{
Repositions columns by name.
}
\examples{
movecols(names(df), "a first; b last; e before c") # assuming a, b and c are column names
movecols()
}
\keyword{columns}
\keyword{reposition}
