% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_omit_col.R
\name{na_omit_col}
\alias{na_omit_col}
\title{Remove rows based on NA values in one column}
\usage{
na_omit_col(data, desiredCols)
}
\arguments{
\item{data}{the data frame}

\item{desiredCols}{the heading of the specific column}
}
\description{
This function removes rows from a dataframe when one column contains NAs
}
\examples{
df1 <- na_omit_col(df, "ID")
na_omit_col()
}
\keyword{column}
\keyword{in}
\keyword{na}
\keyword{remove}
\keyword{rows}
\keyword{with}
