% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OS2EN.R
\name{OS2EN}
\alias{OS2EN}
\title{OS Grid Ref to Eastings and Northings Function}
\usage{
OS2EN(gridreference)
}
\arguments{
\item{gridreference}{The grid reference to be converted. Can be any resolution.}
}
\description{
This function creates a data frame containing the easting and northing coordinates for the SW corner of the grid reference.
}
\examples{
OS2EN()
}
\keyword{Grid}
\keyword{references}
