% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.excel.R
\name{read.excel}
\alias{read.excel}
\title{Reads in .csv, .xls amd .xlsx files to a data frame}
\usage{
read.excel(file)
}
\arguments{
\item{file}{The file path, use file.choose()}
}
\description{
This function will read in any excel file in .csv, .xls or .xlsx.
}
\examples{
read.excel()
}
\keyword{Import}
\keyword{spreadsheet}
