% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_centroid_within_poly.R
\name{st_centroid_within_poly}
\alias{st_centroid_within_poly}
\title{Function for finding a centroid and ensuring that it is within the polygon for irregularly shaped polygons
This function returns an sf point object.}
\usage{
st_centroid_within_poly(poly)
}
\arguments{
\item{poly}{the  sf object containing polygons that you want centroids for}
}
\description{
Function for finding a centroid and ensuring that it is within the polygon for irregularly shaped polygons
This function returns an sf point object.
}
\examples{
sites_lab <- st_centroid_within_poly(sites_crop) 

}
\keyword{centroid,}
\keyword{polygon}
\keyword{sf,}
\keyword{within}
