% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_dot.R
\name{st_dot}
\alias{st_dot}
\title{Function for creating an sf point object to resemble dotting within a polygon for plotting on a map.
This function returns an sf point object.
Eventually the package {ggpattern} will render this function useless, but the functionality is not yet there.}
\usage{
st_dot(x, custom_bbox = NULL, den = 4000)
}
\arguments{
\item{x}{the sf object containing polygons that you want dotted.}

\item{custom_bbox}{the option for a custom bounding box to standardise density in a repetitive mapping process. If blank, the bbox of the sf object will be used.}

\item{den}{the density of dots.}
}
\description{
Function for creating an sf point object to resemble dotting within a polygon for plotting on a map.
This function returns an sf point object.
Eventually the package {ggpattern} will render this function useless, but the functionality is not yet there.
}
\examples{
library(sf)
myshape <- st_read(system.file("data/BerksOxon.shp", package = "ALERC"))
bboxplus <- st_bbox(st_buffer(myshape, 100))
dots <- st_dot(myshape, den = 200, custom_bbox = bboxplus)
library(ggplot2)
ggplot(myshape, aes(col = COUNTYNAME)) +
geom_sf() +
geom_sf(data = dots, aes(col = COUNTYNAME)) +
coord_sf(datum = 27700)
}
\keyword{dot,}
\keyword{ggplot}
\keyword{mapping,}
\keyword{sf,}
\keyword{symbology,}
