% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_min_dist.R
\name{st_min_dist}
\alias{st_min_dist}
\title{Function for creating a numeric vector showing the smallest distance between the
feature in the primary sf object and all features in the second object}
\usage{
st_min_dist(x, y)
}
\arguments{
\item{x}{the primary sf object}

\item{y}{the second sf object}
}
\description{
Function for creating a numeric vector showing the smallest distance between the
feature in the primary sf object and all features in the second object
}
\examples{
library(sf)
records <- readxl::read_excel(system.file("data/records.xlsx", package = "ALERC")) \%>\%
  st_as_sf(coords = c("Easting", "Northing"), crs = 27700)
sssi <- st_read(system.file("data/SSSI.shp", package = "ALERC")) \%>\%
  st_set_crs(27700)
records$DistFromSSSI <- st_min_dist(records, sssi)


}
\keyword{distance,}
\keyword{eucledian}
\keyword{polygons}
\keyword{sf,}
\keyword{spatial,}
