% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_plot.R
\name{st_plot}
\alias{st_plot}
\title{Function for plotting sf objects without having to specific st_geometry
This function plots an sf object.}
\usage{
st_plot(sf, add = F, col = "black")
}
\arguments{
\item{sf}{the  sf object}

\item{add}{default is false, specify add = T if you want the object to be added to the current plot}
}
\description{
Function for plotting sf objects without having to specific st_geometry
This function plots an sf object.
}
\examples{
st_plot()

}
\keyword{plot}
\keyword{sf,}
