% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tverc.R
\name{theme_tverc}
\alias{theme_tverc}
\title{TVERC's custom ggplot theme.}
\usage{
theme_tverc()
}
\description{
TVERC's custom ggplot theme.
}
\examples{
library(sf)
myshape <- st_read(system.file("data/BerksOxon.shp", package = "ALERC"))
ggplot(myshape, aes(col = County)) +
geom_sf() +
coord_sf(datum = 27700) +
theme_tverc()
}
\keyword{ggplot,}
\keyword{theme}
