% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_shapefile.R
\name{write_shapefile}
\alias{write_shapefile}
\title{Function to write out a shapefile with correct .prj}
\usage{
write_shapefile(data, file)
}
\arguments{
\item{data}{The data to be written}

\item{file}{The path and name of the shapefile to be written, without the .shp}
}
\description{
This function creates an sf polygon object, .
}
\examples{
write_shapefile()
}
\keyword{Shapefile}
