% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_png.R
\name{get_png}
\alias{get_png}
\title{Function for reading a png as a raster}
\usage{
get_png(filename)
}
\arguments{
\item{filename}{the  filename of the png file}
}
\description{
Function for reading a png as a raster
}
\examples{
logo <- get_png("G:/Logo.png")

}
\keyword{ggplot}
\keyword{mapping,}
\keyword{png,}
